--- src/main/java/kitchenpos/orders/infra/JpaOrderTableRepository.java
@@ -1,9 +1,0 @@
-package kitchenpos.orders.infra;
-
-import kitchenpos.orders.domain.OrderTable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
-}
