+++ src/main/java/kitchenpos/orders/ui/OrderTableRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.orders.ui;
 
-import kitchenpos.orders.application.OrderTableService;
-import kitchenpos.orders.domain.OrderTable;
+import kitchenpos.orders.eatin.service.OrderTableService;
+import kitchenpos.orders.eatin.domain.OrderTable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
